+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes between versions
 
+## Not yet released
+
+* Fix return typehint on Doctrine entity getters
+
 ## 0.1.1 (2017-05-10)
 
 * Force Doctrine column names in yaml mapping